feat:增加实时统计通道手续费,验证次数

+++ src/main/java/com/rfchina/wallet/server/service/CacheService.java
@@ -101,10 +101,10 @@ public class CacheService {
 	}
 
 
-	public Long getStatisticsYunstVerify(String methodName) {
+	public int getStatisticsYunstVerify(String methodName) {
 		String key = CACHE_KEY_YUNST_VERIFY + ":methodName=" + methodName + ":date=" + DateUtil
 			.formatDate(new Date(), "yyyy-MM");
-		return (long)Optional.ofNullable(redisTemplate.boundValueOps(key).get()).orElse(0L);
+		return (int)Optional.ofNullable(redisTemplate.boundValueOps(key).get()).orElse(0L);
 	}
 
 }
